Test refactorerade till enhetstester och Integrations tester. Test för service bör läggas till
